Query string auth (#77)
* first version of Qs auth

* tests yo
